requirements.txt
tweepy
apscheduler
python-dotenv
openai
tzdata